delete a few tests
